requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==3.2.8
djangorestframework==3.12.4
drf-spectacular-sidecar==2021.9.27
drf-spectacular[sidecar]==0.20.1
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.0.1; python_version >= '3.7'
pillow==8.3.2
psycopg2-binary==2.9.1
pyrsistent==0.18.0; python_version >= '3.6'
pytz==2021.3
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
